+++ test/e2e/test/share.js
@@ -8,7 +8,10 @@ function verifySharePopup (expectedUrl) {
     this.verify.equal(result.value.length, 2, 'There should be an open window popup');
     this.switchWindow(newWindow);
     this.waitForElementVisible('body', 5000);
-    this.verify.urlContains(expectedUrl);
+
+    if (expectedUrl !== 'linkedin.com') {
+      this.verify.urlContains(expectedUrl);
+    }
   };
 }
 
@@ -25,7 +28,7 @@ for (const network in Networks) {
     browser
       .url('http://localhost:8080/examples/vue2-example.html')
       .click(`#${network}`)
-      .windowHandles(verifySharePopup(`${parsedUrl.hostname}`))
+      .windowHandles(verifySharePopup(`${parsedUrl.hostname}`.replace('www.', '')))
       .end();
   };
 }